serve http
Create a streaming HTTP MCP server which can be extended with more
options and configuration then run as a service on the dev's machine.

Plan is to set up local paths to be scanned for modules and/or restricted
to certain file trees.  Especially for separating private/work modules
from public repos or keep projects from pulling in code from other projects
that should not be sharing code.